Precompute inputs to map
Recently when adding support for byte arrays in procedures et al we
introduced a mapping of the input to the procedure so that we can
to byte arrays as necessary. However this means that we will attempt
to map all inputs, byte arrays or not which introduced a regression.
This PR precomputes the inputs that needs mapping and updates only
those.