Coordinates instantiate/close in spatial/temporal indexes
With one-to-one native indexes like number or string,
updates or interactions after call to close() is prohibited
and coordinated inside GBPTree. With the spatial/temporal index
structure, which has lazy instantiation of sub-indexes
there was no such coordination. This meant that a call to close()
could race with an instantiation of one or more sub-indexes.
This in turn would result in opened sub-indexes (GBPTree instances)
that would never get closed.

Contractually this is very good to tighten up so that they all
behave the same. Practically this was found in a population scenario
where an index population was cancelled while running.