Half-deprecate the catchup client timeout
Since 76bacc294171bfd61b02492478218bd3e0fed9ce we longer expect
timeouts to be the means of discovering failed network connections.
This commit does not deprecate the functionality completely, since
it might still come in handy, but sets it at a very high level that
is unlikely to be too eager. A proper deprecation might come with
the next release.