Add a `leader-only` upstream data selection strategy.
We use it in PageCacheWarmupCcIT to ensure that the read replica test copies its files from the core we actually wrote our test data to.
Without this setting, the read replica would copy profiles from any random core, and that core might not have loaded the same amount of data into memory because it has never run any queries.