BuiltInProcedures for index creation
BuiltInProcedures
- db.createIndex
EnterpriseBuiltInProcedure
- db.createNodeKey
- db.createUniquePropertyConstraint

Procedures allow user to specify index provider upon index
and constraint creation. This is otherwise controlled by
dbms.index.default_schema_provider and to change this you
need to restart db.

Like other procedures interacting with schema indexes it takes
schema pattern ":Label( prop1, prop2,... )" as argument as well as
index provider name.
Allowed index provider names correspond to IndexProvider.Descriptor.name()
and allowed values for setting (dbms.index.default_schema_provider).

If label or property tokens are missing, they will be created.

* Why do we need this? *
We need a way to mitigate for the limitations in native string index
(4k length and limited support for contains and ends with).
When creating an index (or constraint) through Cypher the default
index provider is always used. The default provider is controlled with
db.index.default_schema_provider setting. To change this setting user
have to restart the db. This is sad.
Those procedures allow user to specify index provider per index creation
without having to do any db restart.
This gives user two different ways to create an index which is bad and
optimal way would be to have support for this in cypher language. For
3.4 release, that is not feasible and so falling back to procedures is
the way to go.