Add batching APIs to the TokenHolder interface.
This allows the clients of the TokenHolder to query and/or create more than one token at a time.
The IsolatedTransactionTokenCreator is then able to create all the missing tokens in a single implicit transaction,
instead of one transaction per token.

The next step is to expose this API further up the stack, in TokenWrite.