Prevent the CheckPointer from blocking CausalCluster Catchup.
By making the checkpointer not hold on to the PageCache monitor lock
for the whole check point process.

It turned out that this behaviour could inconvenience the Causal Cluster
catchup process.
The CC catchup would ask the page cache warmer to list its files.
To do that, the page cache warmer would call `listExistingMappings` on
the page cache, and that method would also require the PageCache lock.

The fix is that the checkpointer now also uses `listExistingMappings`
to grab a snapshot of the mapped files, and then flush them without
holding on to the lock on the entire page cache.
The `listExistingMappings` method does still take the page cache lock,
but it returns much, much faster than a complete page cache flush would.