Configurable HSTS in server
The HTTP Strict-Transport-Security (HSTS) response header tells
browsers that a webpage should only be accessed using HTTPS instead
of HTTP. Webpage is thus less vulnerable to man-in-the-middle attacks.

This commit adds a setting to configure value of the
`Strict-Transport-Security` response header for neo4j server. Setting
can be used to make neo4j browser endpoint send the HSTS header with
specified max-age, when accessed through HTTPS.

See
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Strict-Transport-Security
for more details.