Configurable HPKP in server
The HTTP Public Key Pinning (HPKP) is a security feature that tells
browsers to associate specified public keys with the webpage. Such
webpage is less vulnerable to man-in-the-middle attacks.

This commit adds a setting to configure value of the
`Public-Key-Pins` response header for neo4j server. Setting
can be used to make neo4j browser endpoint send the HPKP header with
specified public key fingerprints, max-age and other directives when
accessed through HTTPS.

See
https://developer.mozilla.org/en-US/docs/Web/HTTP/Public_Key_Pinning
for more details.