Make `listExistingMappings` not throw eviction exception.
The `listExistingMappings` only _really_ needs to assert that the page
cache has not been closed. It is not strictly necessary for this method
to assert that the page cache is healthy. Furthermore, now that it is
used for implementing `flushAndForce`, checking for the eviction
exception is outright harmful. The `flushAndForce` method is one of
a few ways to clear the eviction exception.

The background eviction thread now also no longer spins forever, if it
cannot evict any pages because they throw exceptions. This makes testing
it easier.