The runtime interface.
Initial draft of the runtime interface. This interface defines
How a query is transformed from a logical plan to a executable
and cacheable query, and how executing that query will result in
a query execution.

In reverse to previous result handling, this interface uses a
push model where the runtime code controls the pace of results.
To enable this, execution clients now have to supply a ResultBuffer
which the query execution will write results to.