Merges StorageStatement into StoreReadLayer
They were split so that StorageEngine#storeReadLayer()
was a getter for a singleton instance and it could create a statement
which contained accessors that required some sort of state
in the reader. This was weird in some ways e.g. where some
StoreReadLayer had to pass in StorageStatement into its methods.
It was a bit schizophrenic. And all practical uses of StoreReadLayer
also required a statement so they always went hand in hand anyway.

Now StorageEngine#storeReadLayer() instantiates a new StoreReadLayer,
which is kept in KTI, just as it would previously be, except now
KTI will not additionally create a new statement too.