Abort replication on missing leader or leader switch
When there is a lot of data queued up behind the replicator, it makes sense
to make a sweeping abort of it all when there is a leader switch or a leader
is missing altogether. A leader switch generally implies, for transactional
work, that the work has to be thrown away anyway due to the switch of lock token.

To keep the safety aspect of exactly-once semantics, we cannot abort once
replication has been attempted, but the replication limit of the Throttler
limits the amount of in-flight replication data which must be finalized.