Made canBeBootstrapped Logic more deterministic
In the Shared discovery service, canBeBootstrapped was calculated for each member by iterating through listening clients
and performing the standard isFirstAppropriate logic. However the ordering was *not* insertion based, as with Hz, but
instead based upon the uuid of the clients. This was causing issues where two clients would attempt to bootstrap in rare
cases, leading to test flakiness.