Moves location of template for the source file header text
The template for all source headers, which includes the
copyright notice and licensing, is now located under the
build/ top level directory, alongside the checkstyle
config.

poms have been updated to point to the correct relative path
for each module, moving away from the classpath inclusion that
was used before.