Store only minimal splitter in internal nodes in GBPTree
Keys in internal nodes only act as guide posts for search and thus we
only need to store just enough to separate the sub ranges from each
other.

For keys with dynamic layout such as strings this can be used to save
space within internal nodes and instead of storing full string we may
only need to store the first few characters depending on the nature
of the strings.