Retry backups in replace member stress test
The backup server is not available when doing store copies for
example so a robust solution has to retry over a period of time.

This also adds a basic assertion on the test as a whole, that it
has to do at least a few replacements to be considered a pass.