Log full message on protocol error in Bolt
This commit makes `MessageDecoder` log full content of the message
buffer if it fails to decode a message from this buffer. Hex dump of
the buffer is logged to the internal log. Network connection is then
closed because of the error.

Also improved `BoltRequestMessage` enum to better handle illegal
message signatures and not use exceptions for flow control.