Skip message processing when Bolt connection terminated
Bolt connection can be terminated asynchronously by the Netty's
`#channelInactive()` callback. This will mark the Bolt state machine
for termination. When worker thread detects termination it rolls back
the open transaction and clears the existing result handle.

Previously, it also tried to continue streaming the result after
processing the termination signal. This caused an attempt to stream
a `null` result and an NPE. This NPE only appeared in the database
logs. It wasn't possible to stream it back in a FAILURE message
because the connection has been terminated.

This commit fixes the problem by making Bolt state machine skip message
processing when it is terminated. It also makes `ErrorReporter` log
full status code with classification.