Don't load property keys if label has no indexes
When many, many labels are being added to many nodes
and those labels don't have indexes, loading property keys
is unnecessary and can drastically slow down queries if
the property keys aren't already in the page cache.

In some cases this can turn a query that should take a few
minutes into one that takes an hour or so.