Restore fallback to interpreted for slotted runtime compilation problems
Slotted is supposed to support all queries, but there is a bug in
slot allocation of FOREACH which was hidden by fallback to interpreted.
When this fallback was removed, this bug was discovered in a downstreams
build. To temporarily fix the downstream issue, we're reinstating the
fallback in interpreted for now, but we will soon come back to fix the
actual bug.