Refactor 3.3 and 3.5 cypher compilation to use a composite compiler
This way, the versioned parts of cypher are hidden behind the
new CypherPlanner interface for all versions where we provide
compatiblity on the logical plan level, while runtimes are easily
shared for all these versions. In addition more step are taken toward
moving functionality which is not runtime dependent out of the runtime
code.