Fix: revert to ignoring erroneous parameter values if they are not used
Core API users used to be allowed to pass in arbitrary objects as
parameters. If these were not used by Cypher, we would silently ignore
them. From 3.4.0 this was changed to always fail the query if any
parameter was not convertible to a cypher type. This commits reverts
to the previous lenient behaviour, as this change caused problems for
some users.