Client side hostname verification with CN and SAN
How hostname verification works:
server public certificate includes an extension for Subject Alternative
Names, which is a list of matchers (such as *.neo4j.com).

the client, after obtaining the servers public certificate and
negotiating TLS handshake, checks the hostname it used to connect
against the Common Name & Subject Alternative Names declared in the
servers public certificate.

If the address used doesn't match the server CN or SANs then the
connection is terminated.

Note: The server side does not do hostname verification, though by
having Subject Name Identification enabled in the protocol it may be
possible, but so far there hasn't been such a need.

Changes in this PR:
introduced OnConnectSslHandlerInjectorHandler that replaces itself with
a netty SslHandler once a connection is established. SslHandler
(SslEngine really) needs the hostname used for the connection during
init, so we create the correct handler with
OnConnectSslHandlerInjectorHandler.

created TrustManagerFactoryProvider, moving implementation out of
SslPolicyLoader though not necessary for hostname verification

legacy SslPolicies do not have hostname verification