Refactor bolt state machine to be more extensible
Bolt state machine is responsible for keeping track of the
corresponding bolt connection and processes incoming bolt messages.
It is not extensible because there is no clear set of state machine
interfaces defined and states are represented as an enum. Bolt
connection abstraction, which is supposed to independent of the bolt
state machine, always uses the version 1 (and only) state machine.

This commit introduces a `BoltStateMachine` which attempts to be
protocol version independent. Instead of exposing methods for all
protocol messages it has a single method that takes a generic message
object `StateMachineMessage` as an argument. States of the state
machine are now represented as classes which can process a generic
incoming message and return the new state.

Also extracted set of potentially common interfaces and classes from
`org.neo4j.bolt.v1.runtime` to `org.neo4j.bolt.runtime` so that they
can be reused for different versions of the state machine.