RETURN true AND [1,2,3] -> [1,2,3]?
For queries like `RETURN true AND [1,2,3]` we do two rewrites that causes
the query to return `[1,2, 3]` instead of `true`.

We start with `Ands(TRUE, list)` and then in the first step we remove
all `TRUE` and in the second step we check if `ANDS` contains a single
element  and if so we remove the `ANDS` altogether. However this last step
is not safe since we will now forget that the list should be coerced to a
boolean value.

In 3.5 we will solve this by doing the coercing to boolean on the AST
instead of when creating runtime expressions.