Change SecurityModule dependencies and life cycle
Instead of having the SecurityModule take the platform LifeSupport
as a dependency and adding its LifeCycle components to it from inside,
it now has its own LifeSupport and LifeCycle, and the edition module
that loads it is responsible for adding it to the platform LifeSupport.

Also expose AuthManager and UserManagerSupplier directly in
EditionModule so they can be accessed without having to use the
dependency resolver.