Constraint index violation won't fail BatchInserter#shutdown()
In BatchInserter the label index and created indexes are populated
when calling shutdown(). Previously this was done single-threaded
and with some home-grown code that was similar to what
IndexingService would do and would also fail shutdown(), i.e.
throw exception from it if there was an index entry conflict
in a uniqueness index. This would leave the database, and indexes
especially in an arbitrarily undefined state and would be
confusing to user making the insert. Was the insert successful
or not, should it be performed from scratch again?

This has been changed to instead instantiate an IndexingService
and let it populate the indexes in a multi-threaded fashion.
Any failed indexes and would neither affect other indexes nor
shutdown() itself. Such failed indexes will instead simply be
marked as FAILED and user can decide what to do about it.

Fixes #10738