Disable prometheus during backup to fix port conflict issues
Due to backup being a neo4j instance it tries to load all kernel extensions.
Since metrics is an extension it will load it, and if prometheus is enabled,
bind it to a network port. If the server you are doing a backup runs on the
same machine backup will thus fail with a `BindException`.

By explicit disable prometheus during backup this issues is resolved.

Current workaround is to have a custom backup `neo4j.conf` where prometheus
is disabled and providing the config to the backup command with
`--additional-config` argument.