Do not trim stack trace during test execution output
Surefire trimStackTrace is a flag that controls
whether to trim the stack trace in the reports to just the lines
within the test or show the full trace. The default value of this flag is: true.
Default value does not suit us because stack traces that tests throw on CI
are basically lost with that option enabled.