Do not log multiple failures for one Bolt channel
All exceptions in an established Bolt channel are handled by the
`HouseKeeper`. It takes care of both inbound and outbound errors.
Outbound errors are redirected to the channel pipeline because
`ChunkedOutput` uses a void channel promise for every `#writeAndFlush()`
call. This promise performs the redirect to the pipeline.

Every write-and-flush operation is executed by the event loop thread.
Multiple such operations might be queued if the even loop thread is
delayed or busy. All of them will fail with the same error if event
loop group is shutdown. It might get shutdown by the Bolt server during
a database shutdown. Such situation made `HoseKeeper` print a lot of
repeated error messages to the log. Every message came from a terminated
write-and-flush operation.

This commit makes `HouseKeeper` print nothing when error happens during
the event loop shutdown. It will also only print a single error per
channel. This should be enough since the channel is closed in
`#exceptionCaught()`.