Large refactor of execution result
PLAN DESCRIPTIONS:

Plan descriptions are now generated directly from the logical plan,
instead of being based on the physical operators. This makes the
plan descriptions for all runtimes be the same, appart from any added
arguments.

To enable profiling, runtimes now produce a `QueryProfile` if the plan
is executed with profiling turned on. This query profile is then handed
to the plan description builder to enrich the plan description.

INTERNAL EXECUTION RESULT:

The InternalExecutionResult has been cleaned of test-only code - that
functionality now lives in RewindableExecutionResult. Also, things have
changed so that StandardInternadExecutionResult now is it's on class
responsible for execution meta data and policies around materializing
results early.

ClosingExecutionResult is responsible for all closing of the execution
results due to end-of-rows or exceptions.

Runtimes implement a much slimmer RuntimeResult instead of
InternalExecutionResult. This only holds the primary functionality
related to getting results and profile information. Notably no runtime
is any longer doing any custom closing, results materialization, or
value conversion.