Update script to better align with different versions of PS
cmd.exe is a bit picky on handling double quotes on the command line arguments passed over and somehow behaves differently on different versions of Windows/Powershell.

This fix is done based on the observation on Windows 7 and verified that  the new code works both on Windows 7 (PS 2.0) and Windows 10 (PS 5.1).