Await for database startup before try to restart slave services
Ha mode switcher defined in ha module and placed in life
before some of the database components. Its possible that voting and
even switching to a new state will happen before all of the components of
platform life will gonna finish its start.
To be sure that we stopping services (on slave switch)
that are already started and we do not race anywhere
we will gonna wait till database will be available before doing that.