Fix a non-deterministic bloom fulltext index bug.
There was a data race where a fulltext index would come online, but a query using that index would observe the index as missing, which causes an exception to be thrown.
The fix is to simply remove the check from the FulltextIndexProvider and instead rely on the kernel to do it for us. Which it does. Except, it does it correctly and without any races.
There was also a second problem, where the exception thrown above would be ignored by the bloom query procedures, and it would then look like the index was there but empty.
This ignoring of the exception meant that it took far, far longer to debug these problems.