Add setting unsupported.dbms.tx_log.fail_on_corrupted_log_files
This settings is true by default and will tell the database to shutdown
if any errors in the transaction logs are encountered. If set to false,
the database will try to recover as much as possible and remove the
corrupted part of the transaction log. This might leave the database in
an inconsistent state and is therefor and opt-in and not on by default.