Reduce memory usage when replicating transaction
TransactionRepresentation are now serialized by netty. Previously it
would be serialized in ReplicatedTransactionFactory before it was sent
 to the raft replicator. This operation would
write the entire transaction to a heap buffer and then copy that buffer
to an array which would use 3 times the tx size.

Now the TransactionRepresentation is handled in
TxRepresentationMarshal. This will write one storage command at the time
directly to nettys direct buffer. Hence no extra heap is used
(1 time tx size instead of 3 times).

Previous implementation would also write the entire transaction to
nettys direct buffer. Because CommandStream is now an Iterable it will
write one command at the the time which is then sent in chunks.
This means that the memory allocated in netty will only be as large as
the larges StoreCommand in the transaction, before it would be the size
of the entire transaction.

To avoid allocating direct buffers that are larger than nettys pooled
chunks a BoundedNetworkChannel is used which will write chunked buffers
to a queue. No buffer will ever be larger than the provided max size.