Use Fallback to interpreted if slotted finds PatternExpression
Slotted runtime cannot properly handle PatternExpressions or
PatternComprehensions. In 3.4 we solve this by falling back to
interpreted.

In 3.5 the planner will plan these Expressions away, but the code
to fallback should be kept in case we forgot a scenario where
the planner does not plan such expressions away. The
FallbackAcceptanceTest is useless in 3.5 though, so do not forward merge
it.