Implementation of minimalSplitter in GenericKeyState
For text, minimal splitter is the subpart of right keys byte array up to
the first position on which left and right key differ from each other.
This is the same way that native string index implement minimalSplitter.

For arrays, works the same way, but instead of comparing byte by byte
we compare one whole position at the time.

For array of text, there is an opportunity to also compare on the individual
string level on the position on which the two keys first differ. This is left
as a todo for now.