Set `needsValue` conditionally in NodeValueCursor
The `DefaultNodeValueIndexCursor` would alwazs (when possible)
set `needsValue` to true and deserialize values even if the caller
of the index operation did not need the property values.

Now, the caller (e.g. NodeIndexSeeker) decides whether values are
needed.

Adds parametrized tests in `NodeTransactionStateTestBase` to hit
both code paths when values and needed and when not.