Use case classes for IndexedNodeWithProperties
The case classes IndexedNodeWithProperties and
IndexedPrimitiveNodeWithProperties replace previously used tuples.
This makes it easier to change the definition in one place later one.

Also the signatures of runtime stuff has been changed to use Arrays
instead of Seqs where possible, since we know lengths ahead of time.

This commit also fixes a problem in lockingUniqueIndexSeek where
it tried to get the values from the index, which is not allowed
for ExactPredicates.