Avoid ConcurrentModificationException in the SchemaCacheState copy-constructor.
The collections were all copied, but some of them had other collections as their values, and these were not copied, but instead directly transferred to the new SchemaCacheState copy.
This could result in ConcurrentModificationException on racy accesses to the SchemaCacheState.
To fix this, we now also produce copies of all of those inner-collection values.