Set GetValueFromIndex in logical leaf planning
AbstractIndexSeekLeafPlanner and indexScanLeafPlanner now query
the IndexDescriptor for the ValueCapability of an index, given
the CypherType of the Seekable Expression.

The set the behavior to `GetValue` if the index is capable,
regardless of whether the value is needed in the actual query.

This commit includes test support to mock indexes with values.