ProvidedOrder on index plans and sort. Do not sort already sorted plans.
Each LogicalPlan must now define a ProvidedOrder. This is at the moment
only implemented in index operators and Sort. The steps that plan
index operators set the ProvidedOrder according to the capability of
the index. The sortSkipAndLimit checks if the ProvidedOrder of the
current plan satisfies the RequiredOrder of the PlannerQuery before
planning a Sort operator.

(Does probably not compile)