Make ProvidedOrder an Attribute. Gather attributes in Context.
ProvidedOrder was in the previous few commits set directly on the
logical plan. This does not work as soon as other attributes (e.g.
cardinality) play a role in deciding the ProvidedOrder. Therefore
it was made an attribute.

As part of the work making Solved and Cardinality attributes, the
signatures of basically the whole planner were changed to include
these two arguments. Instead of adding a third onw now, this was
reverted and all PlanningAttributes are gathered in the
LogicalPlanningContext.

Small changes to LogicalPlanningTestSupport to test indexes with values
and indexes with ordering.