Introduce explicit handling on cached node properties
This allows us to defend against misalignment between the in row property
values and any property write which could occur during the query. Special
runtime handling of cached node properties let's us check whether the
property for that node has been changed in this transaction, and in that
case use the tx-state value instead of the in-row value.