Label index reader can be kept up2date with concurrent updates
This is important for index population to see created entities
right in front of it while it's scanning and pauses to apply
external updates. Internally each time a batch of updates
have been applied the label index reader is restarted from the
current idRange and seen ids within that idRange skipped.

This allows populators to have the guarantee that updates coming
in from the scan haven't been indexed before and can therefore
be indexed as efficiently as possible.