Disable shrinking in test
SemanticIndexAcceptanceTest is yet again flaky - it started just
a week ago and I suspect there is an actual bug hiding here somewhere.
However the shrinking done by scalacheck here hides the actual error
which makes it even trickier to investigate. This PR just disables the
shrinking in order to make the investigation easier.