Shutdown custom backup service page cache and scheduler
In cases where BackupProtocolService was created without externally provided
pageCache another separate independent instance was created and it was need closed.
In addition to leaked page cache memory, the scheduler that independent page cache
was using was also never closed and that was the source of leaking threads.

To be able to avoid memory and thread leaks new
BackupProtocolServiceFactory was introduced to allow constructions of
various variants of BackupProtocolService.
Also service class itself updated to be AutoCloseable.
As result now BackupProtocolService clients have the ability to release all
resources by closing service after required operations.