Support index values for WHERE n.prop <= 42 AND n.prop % 2 = 0
In the case where one predicate is solved by an index and another
predicate also uses the same property, it now does not go to
the property store but uses the index provided values instead.

Changes `ReplacePropertyLookupsWithVariables` to be a full-fledged
rewriter that happens after planning, instead of calling into
it in many different places from the planner (and remove corresponding
unit tests).