SpatialIndexKey no longer generates number-values
When calling SpatialIndexKey#asValue() it would generate a NumberValue
with the idea that it would aid sampling somehow.
Sampling for native indexes has been rewritten to no longer
look at materialized values so this no longer makes sense.
This also makes it possible to remove SpatialHitIndexProgressor,
which would otherwise be necessary to not have NodeValueClientFilter
filter out spatial values when `needsValues` was `true`.