Avoid logging to debug log twice in CoreMonitor
SimpleLogService, which the core monitor gets the log providers from,
already wraps the user and debug log in a DualLogProvider. This means
that the user log will always log to debug log as well. When the
CoreMonitor again wrapped the two providers in a DualLogProvider,
many of the messages would end up being duplicated in the debug log.