Change order of reporting page cache events and processed in ParallelPageLoader
A processed counter is used on close in ParallelPageLoader itself
to know when we are done.
Reported page cache events are used heavily in tests to verify page cache
results of different sorts.
While the final eventual result will be the same after some moments
it's better to report page cache events before incrementing processed
counters for 2 reasons:
1. When we closing page loader all its work should be already done.
2. Reporting events before we are closed makes verification of page
   cache events predictable.