Do not use BigDecimal in selectivity estimation
We used to use BigDecimal in around half of the calculation,
then methods Selectivity (which use doubles) in the rest.

You only get real benefit of using BigDecimal if you use it all the way
before producing a resuolt (in this case deciding on the cardinality
of a LogicalPlan). If you compute certain parts with BigDecimals,
but then convert to double before doing more calculations, that
increased precision is likely gone.

Therefore,and since selectivity estimation is guesswork anyway,
I removed BigDecimal calculations altogether.