Improved on failure/archive functions in GenericNativeIndexProvider
FusionIndexProvider has an archive function for failed indexes
as they are rebuilt. Now GenericNativeIndexProvider uses the
same function.

GenericNativeIndexProvider could previously fail on instantiating
its Layout, even after it was concluded that the index was broken
and had to be rebuilt and so would fail the startup of the database
as a whole. Now this has been fixed so that the broken tree
cannot interfere with rebuilding of broken indexes.