Relax JVM version check in startup scripts
Allow VM that has a version string different from 1.8 to allow to start
neo4j on latest JDK.
The restriction that will not allow starting neo4j on older JVM version
is still in place.
Also, update scripts to indicate that they are actually generated based
on the template. Align template and fixes that were only applied to scripts.
Add tests that verify windows startup scripts.