Add the concept of transaction data _revisions_ to the transaction state.
This is useful for auxiliary transaction state providers, because it allows them to check if any data has been changed in a transaction since they last updated their auxiliary transaction state.
The transaction data revision is updated every time there are any data changes in a transaction.
The transaction revision is otherwise an opaque number, except the value of zero means that the transaction has not seen any data changes since it was started.