Make sure we stream transactions as far as is required or fail
This is a rare race which can occur when transaction log pruning
comes in while an iteration over transactions spanning several
files is in progress. If the pruner deletes files ahead of us
then the kernel transaction cursor will not notice, which
previously would lead us to fall short. A safety check is now
added which works regardless of cursor behaviour.

This also adds a sanity check on the expected sequence of
the streamed transactions.