Skips coordinate data in internal point keys
Since internal keys will never be used to materialize values
and in the case of points the overhead of keeping coordinates
in the internal keys is much more visible due to a single range query
resulting in lots of sub-range queries. Therefore the internal nodes
are traversed much more than for other types.

State for internal keys and state from derived point of a sub-range
looks the same, so checks for writing points w/o coordinates have been
removed, which is essential for allowing internal keys to have no coordinates.
It is now therefore a bigger reeponsilitity of the developers to not
write keys w/o coordinates to leaf nodes.