Detached database recovery.
Separate database recovery from database server lifecycle and
allows particular database recovery without starting full database
service. Introduces new facade `Recovery` entity that can be used for
recovery or to check is database requires recovery.
Recovery performed in a completely new separate (from the database
it needs to recover) lifecycle.
New separate instances of kernel extensions will be started to
participate in recovery. Only extensions that marked with
`@RecoveryExtension` will be instantiated for this purpose.
At the end of recovery new checkpoint will gonna be performed.
As before recovery still does not guarantee that indexes are fully rebuilt.
In case if the database does not require recovery,
performing recovery on top of it will not gonna do anything.