Do not force store file snapshot on udc approximate store size calculation
Before this PR each UDC report of approximate store size and that caused
snapshoting of all the indexes in the store with by itself can cause
quite a load in case if you have lots of them, also provoking locking
issues with lucene.
This PR will change that behaviour to simply count database directory
size instead.