Transaction kill procedures.
Introduce procedures that allow killing one or several transactions.
As a result of the kill procedure invocation transaction will be
marked as terminated and thread executing marked transaction is
responsible for transaction closing.
On a moment when the transaction is terminated, it will release all its
locks and should not block any other transaction after that.

In case if executing transaction thread will never come back the transaction
will hang in a list of transactions as terminated and we will not perform
any force close.