Fixes a concurrency issue with ParallelNativeIndexPopulator close/drop
Reason being that a drop() call can come in directly to the
specific populator, in addition to the otherwise normal route via
MultipleIndexPopulator, which is overseeing the lifecycle of
individual populators.

Now the close()/drop()/newPartPopulator() calls are synchronized
as to be able to sort this out nicely. close()/drop() are called
once per index population and newPartPopulator is called once
per thread per index population so there's no synchronization overhead
in this.

ConcurrentCreateDropIndexIT could expose this problem before this fix.