New simpler and faster consistency checker
A complete rewrite of the checker code, but keeping many of the
algorithms of the current checker. New checker have none of the
unnecessary and complicated abstractions of the current checker.
More checks exploits the cache for less record reading in general.
Most checks try to take advantage of as many CPUs as possible
and tries to keep I/O sequential. Performance tests indicate an
end-to-end improvement is in the vicinity of up to 10x over the
current checker. The new checker also finds more inconsistencies
than the current checker and can run in more memory constrained
environments.

The general structure is that NewSimpleConsistencyChecker is constructed
and its single #check() method invoked where some individual checkers
are run, like:

- NodeChecker
- IndexChecker
- RelationshipChecker
- RelationshipGroupChecker
- RelationshipChainChecker

The individual checkers run mostly in isolation from each other, but
with one thing connecting them - the CacheAccess. It is basically a
node-centric off-heap chunk of memory taking up 11 bytes per node.
All checkers use this cache, either to write something to
(for another checker to read) or for itself to read later in the store scan.
The checkers are run one after the other. Some checkers also build up
counts state, which is verified against the counts store in the end.

If memory is constrained so that there isn't enough memory to hold
nodeHighId * 11 bytes in memory, the consistency check is run multiple
times, each time for a separate part of the node id range.

The new checker co-exists, for the time being, side-by-side with the
current checker and can be enabled using the setting:

`unsupported.consistency_checker.experimental=true`

And when proven stable will replace the current checker and this setting
removed. Also an internal consistency check tool is added to more easily
run ad-hoc consistency checks from your IDE - o.n.t.c.ConsistencyCheckTool.

Another setting for quickly aborting the check on N found
inconsistencies have been introduced, e.g:

`unsupported.consistency_checker.experimental.fail_fast=10`