Removes dubious CRS settings count validator
This was intended to have been removed previously but wasn't.
The validation was trying to minimize the risk for an index
to get more CRS settings that it could fit into its GBPTree header.
The detection wasn't 100% in the face of concurrency and neo4j
only supports a handful of CRS settings anyway, where the maximum
is in the hundreds.