Add safety count downs to avoid infinite loops
In one SOAK test a count store look-up got stuck in an infinite loop
while trying to lookup an index size. This is possibly some corner case
that can occur on count store rotation, but we did not manage to derive
the conditions which caused it. These changes are added so that we
rather crash the query than never return and effectively loose a thread.