Database level page cache wrapper.
Wrap page cache into separate database level page cache
two restrict and control access to page cache on a database level.
Flush of page cache on database level will flush only files that
related to that database. Lookup of existing mappings will be restricted
to files that belong to that database, etc.
This PR does not change lifecycle of page cache ( we do not flush when we
close database, etc) - that will be done in the separate PR.