Moves StorageEngine API to its own component
SchemaDescriptor and friends also moved to storage-engine-api to replace
the specific IndexDescriptor class. An IndexDescriptor interface was
introduced in storage-engine-api, equivalent to the ConstraintDescriptor
interface. These and some exception classes are shared between
storage engine API and kernel API those specific classes had to move along too
and for this reason kernel-api depends on storage-engine-api, which is better
than previously where the dependency was bidirectional.

Another important change is that IndexReader/LabelScanReader moved
to where the rest of that surface (IndexAccessor, IndexPopulator, IndexUpdater)
etc lives. StorageReader no longer manages these instances either.
This for two reasons:

- A recent addition to IndexReader interface required parts of Kernel API,
  so it couldn't move to storage-engine-api
- It makes much more sense to have the whole index surface live in the
  same place