Stop ongoing indexing jobs before shutdown checkpoint
Shutdown checkpoint is a first thing to happen on DataSource shutdown
and it will try to flush all of the databases files into the storage
device. And it's a good thing to not race with anyone who can create/remove
mapped files during that operation. Also, any operations that can
happen after that can be lost on next start. While it should not be
the case for this particular situation (we should trigger index
rebuilds for broken indexes) its still incorrect and causes unnecessary
races, can cause bugs when we will try to read corrupted index files, etc.