Removes Kernel cursors <--> record storage bit-fiddling dependency
This commit has a lots of good things about it, although I'll start
with the bad thing: Some store cursors init methods now has a
'nodeIsDense' argument passed into them. This isn't necessarily
bad at this point because the StorageNodeCursor exposes it already,
but bad because it's seen as an implementation detail of the
current record storage cursors and should be hidden at some point.
The reason why this argument is needed is only because
relationship/group references can be returned out to a client
which will use them to initialize cursors later on.
Doing that will lose context that the reference otherwise had,
so that context is encoded into the reference by the kernel cursors
and separated out when entering later. An example of this context is
to know whether or not the node it came from is dense.

Regarding the flags... many of these flags were encoded/decoded
when going back and forth between kernel cursors and storage cursors,
communicating subtle things. This created a hard dependency from
the kernel to the current record storage implementation.
Now flags are never encoded into references in normal (non-detached)
use cases, they are instead only encoded into the references
right before leaving the kernel cursor surface. Storage cursors never
encodes or decodes any such flags anymore.

Another weird thing was that RecordStorageTraversalCursor couldn't
filter on type/direction, but relied on the kernel cursors to do this
for it. This has been pulled down into the record storage impl so that
kernel cursors can rely on storage cursors to return the correct things.
For this the storage cursors init methods have been changed to make
type/direction more explicit.