Kernel instantiates indexing PART 1
Moves out instantiation of IndexingService and LabelScanStore to kernel
so that it's no longer the responsibility of the storage engine to do this.
Previously the dependency was weird in that StorageEngine was responsible
for instantiating these services and immediately after kernel would rip
those out of the storage engine using the dependency resolver and use
them, core parts of kernel as they are.