Fix database manager lifecycle in CC
Before this change, lifecycle of the database manager was
essentially managed in two places. It was added to the global
platform LifeSupport and started/stopped manually by core and
read-replica components. This resulted in a correct startup
sequence but incorrect shutdown sequence. Database manager
would shut down before CC components making them attempt to
work on a stopped database and panic.

Problem is fixed by only adding database manager to the global
lifecycle when the database is not in CC mode. CC-specific
components fully manage the database manager otherwise.