Simplify collection daemon to only have states `idle` and `collecting`
This simplifies usage, and make the user experience more allowing. In
particular the actions `collect` and `stop` are now either successful or
idempotent. `clear` simply removes collected data while in the `idle`
state. And it is possible to continue appending data by just restarting
collecting while in the idle state.