Add migrator for online backup address setting
This setting was renamed in 4.0 from 'dbms.backup.address' to
'dbms.backup.listen_address' in order to be consistent with other
listen address settings.

Migrator will replace old setting key (if configured) with the new
one when configuration of the database is constructed. It also prints
a warning saying that a deprecated setting is used.